Type-casting to int

+++ protected/components/Helper.php
@@ -533,7 +533,7 @@ class Helper
      * @return int|string
      */
     public static function return_bytes($val) {
-        $val = trim($val);
+        $val = (int)trim($val);
         $last = strtolower($val[strlen($val)-1]);
         switch($last) {
             case 'g':
